Remove checks on v1 in action creators for samples
