+++ src/__test__/redux/actions/projects/deleteProject.test.js
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
@@ -6,11 +8,11 @@ import deleteProject from 'redux/actions/projects/deleteProject';
 import initialSampleState, { sampleTemplate } from 'redux/reducers/samples/initialState';
 import initialProjectState, { projectTemplate } from 'redux/reducers/projects/initialState';
 
-import { SAMPLES_DELETE_API_V1 } from 'redux/actionTypes/samples';
 import {
   PROJECTS_DELETE, PROJECTS_SAVED, PROJECTS_SAVING, PROJECTS_SET_ACTIVE,
 } from 'redux/actionTypes/projects';
 import { EXPERIMENTS_DELETED } from 'redux/actionTypes/experiments';
+import { SAMPLES_DELETE } from 'redux/actionTypes/samples';
 
 import config from 'config';
 import { api } from 'utils/constants';
@@ -104,19 +106,9 @@ describe('deleteProject action', () => {
 
     // Sets up loading state for saving project
     const actions = store.getActions();
-    expect(actions[0].type).toEqual(PROJECTS_SAVING);
-
-    // Delete experiments
-    expect(actions[1].type).toEqual(EXPERIMENTS_DELETED);
-
-    // Delete sample
-    expect(actions[2].type).toEqual(SAMPLES_DELETE_API_V1);
-
-    // Delete project
-    expect(actions[3].type).toEqual(PROJECTS_DELETE);
-
-    // Resolve loading state
-    expect(actions[4].type).toEqual(PROJECTS_SAVED);
+    expect(_.map(actions, 'type')).toEqual([
+      PROJECTS_SAVING, SAMPLES_DELETE, EXPERIMENTS_DELETED, PROJECTS_DELETE, PROJECTS_SAVED,
+    ]);
   });
 
   it('Dispatches event correctly for multiple samples', async () => {
@@ -125,19 +117,9 @@ describe('deleteProject action', () => {
 
     // Sets up loading state for saving project
     const actions = store.getActions();
-    expect(actions[0].type).toEqual(PROJECTS_SAVING);
-
-    // Delete experiments
-    expect(actions[1].type).toEqual(EXPERIMENTS_DELETED);
-
-    // Delete sample
-    expect(actions[2].type).toEqual(SAMPLES_DELETE_API_V1);
-
-    // Delete project
-    expect(actions[3].type).toEqual(PROJECTS_DELETE);
-
-    // Resolve loading state
-    expect(actions[4].type).toEqual(PROJECTS_SAVED);
+    expect(_.map(actions, 'type')).toEqual([
+      PROJECTS_SAVING, SAMPLES_DELETE, EXPERIMENTS_DELETED, PROJECTS_DELETE, PROJECTS_SAVED,
+    ]);
   });
 
   it('Switches to activeProjectUuid to another project if multiple project exists', async () => {
@@ -146,22 +128,10 @@ describe('deleteProject action', () => {
 
     // Sets up loading state for saving project
     const actions = store.getActions();
-    expect(actions[0].type).toEqual(PROJECTS_SAVING);
-
-    // Switch active proejct
-    expect(actions[1].type).toEqual(PROJECTS_SET_ACTIVE);
-
-    // Delete experiments
-    expect(actions[2].type).toEqual(EXPERIMENTS_DELETED);
-
-    // Delete sample
-    expect(actions[3].type).toEqual(SAMPLES_DELETE_API_V1);
-
-    // Delete project
-    expect(actions[4].type).toEqual(PROJECTS_DELETE);
-
-    // Resolve loading state
-    expect(actions[5].type).toEqual(PROJECTS_SAVED);
+    expect(_.map(actions, 'type')).toEqual([
+      PROJECTS_SAVING, PROJECTS_SET_ACTIVE, SAMPLES_DELETE,
+      EXPERIMENTS_DELETED, PROJECTS_DELETE, PROJECTS_SAVED,
+    ]);
   });
 
   it('Dispatches fetch correctly.', async () => {
@@ -182,7 +152,7 @@ describe('deleteProject action', () => {
     );
   });
 
-  it('Dispatches fetch correctly in api v2', async () => {
+  it('Dispatches fetch correctly', async () => {
     config.currentApiVersion = api.V2;
 
     const response = new Response(JSON.stringify({}));
