+++ src/components/data-management/UploadDetailsModal.jsx
@@ -10,7 +10,7 @@ import { UploadOutlined } from '@ant-design/icons';
 import { useDispatch } from 'react-redux';
 import endUserMessages from 'utils/endUserMessages';
 import handleError from 'utils/http/handleError';
-import { uploadSingleFile, fileObjectToFileRecord } from 'utils/upload/processUpload';
+import { createAndUploadSingleFile, fileObjectToFileRecord } from 'utils/upload/processUpload';
 
 import UploadStatus, { messageForStatus } from 'utils/upload/UploadStatus';
 import downloadSingleFile from 'utils/data-management/downloadSingleFile';
@@ -68,7 +68,7 @@ const UploadDetailsModal = (props) => {
       return;
     }
 
-    uploadSingleFile(newFile, activeProjectUuid, sampleUuid, dispatch);
+    createAndUploadSingleFile(newFile, activeProjectUuid, sampleUuid, dispatch);
     onCancel();
   };
