+++ src/redux/actionTypes/samples.js
@@ -13,12 +13,7 @@ const SAMPLES_UPDATE = `${SAMPLES}/update`;
 /**
  * Delete sample.
  */
-const SAMPLES_DELETE_API_V2 = `${SAMPLES}/deleteApiV2`;
-
-/**
- * Delete sample following apiv1 version, remove when doing the final step of cleaning up v1 code.
- */
-const SAMPLES_DELETE_API_V1 = `${SAMPLES}/deleteApiV1`;
+const SAMPLES_DELETE = `${SAMPLES}/deleteApiV2`;
 
 /**
  * Update files in sample.
@@ -60,8 +55,7 @@ const SAMPLES_LOADING = `${SAMPLES}/loading`;
 export {
   SAMPLES_CREATE,
   SAMPLES_UPDATE,
-  SAMPLES_DELETE_API_V2,
-  SAMPLES_DELETE_API_V1,
+  SAMPLES_DELETE,
   SAMPLES_FILE_UPDATE,
   SAMPLES_LOADED,
   SAMPLES_SAVING,
