+++ src/redux/actions/samples/createSample.js
@@ -14,15 +14,13 @@ import endUserMessages from 'utils/endUserMessages';
 
 import { sampleTemplate } from 'redux/reducers/samples/initialState';
 
-import config from 'config';
-import { api } from 'utils/constants';
 import UploadStatus from 'utils/upload/UploadStatus';
 
 const createSample = (
   projectUuid,
   name,
   type,
-  filesToUploadV2Only,
+  filesToUpload,
 ) => async (dispatch, getState) => {
   const project = getState().projects[projectUuid];
   const createdDate = moment().toISOString();
@@ -36,9 +34,6 @@ const createSample = (
     },
   });
 
-  let url;
-  let body;
-
   const newSample = {
     ..._.cloneDeep(sampleTemplate),
     name,
@@ -51,26 +46,18 @@ const createSample = (
       .reduce((acc, curr) => ({ ...acc, [curr]: DEFAULT_NA }), {}) || {},
   };
 
-  if (config.currentApiVersion === api.V1) {
-    url = `/v1/projects/${projectUuid}/${experimentId}/samples`;
+  const url = `/v2/experiments/${experimentId}/samples/${newSampleUuid}`;
 
-    body = _.clone(newSample);
-  } else if (config.currentApiVersion === api.V2) {
-    url = `/v2/experiments/${experimentId}/samples/${newSampleUuid}`;
-
-    let sampleTechnology;
-    if (type === '10X Chromium') {
-      sampleTechnology = '10x';
-    } else {
-      throw new Error(`Sample technology ${type} is not recognized`);
-    }
-
-    filesToUploadV2Only.forEach((fileName) => {
-      newSample.files[fileName] = { upload: { status: UploadStatus.UPLOADING } };
-    });
-
-    body = { name, sampleTechnology };
+  let sampleTechnology;
+  if (type === '10X Chromium') {
+    sampleTechnology = '10x';
+  } else {
+    throw new Error(`Sample technology ${type} is not recognized`);
   }
+
+  filesToUpload.forEach((fileName) => {
+    newSample.files[fileName] = { upload: { status: UploadStatus.UPLOADING } };
+  });
 
   try {
     await fetchAPI(
@@ -80,7 +67,7 @@ const createSample = (
         headers: {
           'Content-Type': 'application/json',
         },
-        body: JSON.stringify(body),
+        body: JSON.stringify({ name, sampleTechnology }),
       },
     );
