+++ src/redux/actions/samples/index.js
@@ -4,8 +4,8 @@ import updateSampleFile from './updateSampleFile';
 import loadSamples from './loadSamples';
 import deleteSamples from './deleteSamples';
 import saveSamples from './saveSamples';
-import createSampleFileV2 from './createSampleFileV2';
-import updateSampleFileUploadV2 from './updateSampleFileUploadV2';
+import createSampleFile from './createSampleFile';
+import updateSampleFileUpload from './updateSampleFileUpload';
 
 export {
   createSample,
@@ -14,6 +14,6 @@ export {
   loadSamples,
   deleteSamples,
   saveSamples,
-  createSampleFileV2,
-  updateSampleFileUploadV2,
+  createSampleFile,
+  updateSampleFileUpload,
 };
