+++ src/redux/reducers/projects/samplesDelete.js
@@ -4,9 +4,9 @@ import produce from 'immer';
 import initialState from './initialState';
 
 const samplesDelete = produce((draft, action) => {
-  const { projectUuid, sampleUuids } = action.payload;
+  const { projectUuid, sampleIds } = action.payload;
 
-  _.pullAll(draft[projectUuid].samples, sampleUuids);
+  _.pullAll(draft[projectUuid].samples, sampleIds);
 }, initialState);
 
 export default samplesDelete;
