+++ src/redux/reducers/samples/index.js
@@ -1,8 +1,7 @@
 import {
   SAMPLES_CREATE,
   SAMPLES_UPDATE,
-  SAMPLES_DELETE_API_V2,
-  SAMPLES_DELETE_API_V1,
+  SAMPLES_DELETE,
   SAMPLES_FILE_UPDATE,
   SAMPLES_LOADED,
   SAMPLES_SAVING,
@@ -35,11 +34,7 @@ const samplesReducer = (state = initialState, action) => {
       return samplesUpdate(state, action);
     }
 
-    case SAMPLES_DELETE_API_V2: {
-      return samplesDelete(state, action);
-    }
-
-    case SAMPLES_DELETE_API_V1: {
+    case SAMPLES_DELETE: {
       return samplesDelete(state, action);
     }
