+++ src/redux/reducers/samples/samplesDelete.js
@@ -1,10 +1,10 @@
 import _ from 'lodash';
 
 const samplesDelete = (state, action) => {
-  const { sampleUuids } = action.payload;
+  const { sampleIds } = action.payload;
 
   let updatedObject = null;
-  updatedObject = _.omit(state, sampleUuids);
+  updatedObject = _.omit(state, sampleIds);
 
   return updatedObject;
 };
